Split version declarations in shared/version into several files
